Add files via upload
Use of regression analysis, cluster analysis and ANOVA to solve hypothetical problems within phone technology company
Demonstration 1: The illustration below shows the use of multiple linear regression analysis to determine the relationship between sales of Samsung phones (dependent variable) and its corporate success factors such as marketing, technology, and branding (independent variables). 

Demonstration 2: Following the recall of the Samsung note 7 phones, Samsung has resorted to identifying 5 different manufacturers of the lithium batteries and asked that they submit safety thresholds for one (1) batch of batteries produced in the last quarter of 2022 in anticipation for production in the year 2023.
To evaluate the best supplier with the lowest safety threshold per batch anova is used to determine the difference in means to prevent a repeat of the September 2016 error.

Demonstration 3: To determine what combination of employee benefits is currently preferred by different clusters or groups of employees to position the company to tailor these benefits according to employee profile and to channel resources towards this with the emerging economic downtown and the anticipated scramble to retain talent amongst the big tech companies and organisations.
An employee benefit survey data was obtained for the total 347 employees at the headquarters in Seoul.